VCOM-3281 Changes ResourceOperationService to use system context
Changes the ResourceOperationService to use system context to perform calls
to ResourceOperationSpecService so that all resource operations are accessible.

Provisioning Service is storing configuration data like ResourceOperationSpec
without storing tenant data. This data is accessible by the Provisioning
Service when using local PhM because the authorization configuration in
ProvSvc allows access to documents that do not have tenantLinks stored.
When using Symphony to host PhM services it restricts access to such
documents. This causes queries over such data to be empty when executed
with a normal user. These documents can be access with the system user only.

Change-Id: I8360142c7a274b14528682e67fe4b4aaac113f18