Automatically expand tagLinks in all resources
Searching resources by tags is very suboptimal because first
tag links should be found based on the search criteria (and
these could be many if wildcards are used and/or the seach is
case-insensitive), and then the resources containing these links
should be searched.

To improve this the proposal is to automatically expand tagLinks
whenever they are changed on the resource through POST/PUT/PATCH.
Fortunately tags are immutable so keeping tags and tagLinks in sync
is relatively easy and cheap.

The implementation allows to add tag metadata in the future too
(e.g. internal vs. discovered, etc.) but this may not be that
trivial as these 'origins' are not immutable.

Change-Id: I53c468289712e1541ea615a9d3e11acba5b79406