Skip grooming for resources that never had endpoint links
In provisioning we sometimes allocate resources before knowing the
endpoint which they will land on. Sometimes these are "desired state"
documents, sometimes they are allocated states that will be provisioned
shortly.

If resource grooming kicks off for any endpoint while we have resources
with no initialized endpoint links, they will get removed.

This change fixes this in a generic way by introducing a "__noEndpoint"
custom property that is set on creation and removed on the first endpoint
assignment. Once removed, it can never be set to true again for the
same resource.

This is currently implemented for disk states only but may be enabled
for other resources too.

Change-Id: I5073310d76c1d7884c299995524d13fb5f252d5c