[VCOM-3650] Provisioning independent disk fails when storage profile is not defined
In a setup with csp auth enabled and the request is from BP, the
tenantLinks are populated with projects and user as well. Resources like
datastore / storage policy which are enumerated has only the org link in
the tenantLinks and hence the query to fetch the datastores in this
scenario fails.

Fixed by filtering the tenantLinks to contain only the org link for query.

Change-Id: I276c4621d7280904a44fa2029c0e1441d72dbf46