VCOM-3559 Security groups not getting cleaned up after LB is deleted
* Added retry logic to security group removal code in order to prevent
cases where depended resources (machines, load balancers) have not been
fully cleaned yet. For instance, when a load balancer is removed, there
may be a delay between the time the API to delete the LB completed and
the time the LB is actually removed from AWS. This new logic will retry
a number of times until the SG is deleted successfully or the maximum
number of attempts has been reached.
* Added unit test to cover this use case.

Change-Id: I287e35b3df48e43a5f75664c1570e54f91fda3d3