vSphere - Handling persistent disks during VM deletion
When the compute delete is invoked, we issue a call to vsphere to delete
the VM. vSphere deletes all the hard disks associated with the Vm, when
the VM is deleted. If user has configured persistent for certain disks,
then those disks should be retained so that Vm delete doesn't delete them.
This change handles this use case of detaching the persistent disks before
the VM delete is issued to the vsphere and makes this disks as AVAILABLE
in our inventory so that it can be attached to a different VM at a later
point in time.

When the VM is deleted, if the CD-ROM disks of VM has a associated file
backing those things are left in the vsphere even after the vm is deleted.
Have handled this scenario of clean up of directory of CD-ROM associated
to a specific VM.

Added test case for this scenarios.

Change-Id: Idb1f3e65cfc63ee425232be51eb5c5084e423299