VCOM-3771 NSX-T network profile UI doesn't allow selecting opaque subnet for "External subnet"
Issue:
This seems to be a regression due to the changes made for following: https://bellevue-ci.eng.vmware.com:8080/27205
The endpoint filter was changed from "endpointLink" to "endpointLinks.items". The subnetwork "DONT-DELETE-Dev-01" does not have "endpointLinks" property. It still has only "endpointLink" property.

Fix: The subnet state generated from the network ("createSubnetStateFromNetwork") was not setting the newly introduced endpointLinks property. Made a fix the set this property.

Testing Done:
- Create an NSX-T and vSphere endpoints sqa-nsxt21-manager.sqa.local and sqa-nsxt21-vc.sqa.local respectively. Made sure the sub-network enumeration for the DONT-DELETE-Dev-01 has the newly introduced property endpointLinks.
- Created the Network Profile and the DONT-DELETE-Dev-01 is visible in the external subnet dropdown.
- Successfully tested following tests:
   - TestVSphereEnumerationTask

Change-Id: I9e07bce051fbcc2c391b1de2f5e9b36f36418c77