VCOM-3932: [HA] AzureLoadBalancerProvisionIT: Required security rule parameters are missing for security rule
AzureLoadBalancerProvisionIT.testProvision[0:loadbalancer] is failing very
often in HA postfligh job with error CloudException: Required security rule
parameters are missing for security rule.

The root cause is that AzureLoadBalancerService gets LoadBalancerStateExpanded,
which has polymorphic field: Set<ResourceState> targets. When this request is
remote, the Json deserialization instantiates the objects for targets as
objects of ResourceState instead of ComputeState or NetworkInterfaceState. The
result is that LoadBalancerStateExpanded result is not correctly propagated in
AzureLoadBalancerService.getTargetStates.

The fix is to add a custom JsonMapper that is able to deserialize the
polymorphic types as derived objects instead of ResourceState. The custom
mapper is only registered for the types that are used polymorphically.

Photon model clients have to register the custom mapper by explicitly calling
LoadBalancerStateExpanded.registerCustomDeserializer();

Change-Id: I1cb29d6bc03a6461fa171d2835e335d77aabcc80