VSphereVMSnapshotEnumerationHelper.processSnapshot should use vSphere executor to process the results
In VSphereVMSnapshotEnumerationHelper.processSnapshot there is a query to get
the vm snapshots, and for each result, there is a call to vSphere which is run
in the queries executor. However, the call to vSphere is a blocking call, which
blocks the thread in the queries executor (something which should not happen).
Because processSnapshot is run for each vm in the vSphere enumeration, it was
very easy to hold all the threads in the queries executor, holding all the
queries in the host.

The fix is to run the processSnapshot handler in the vSphere executor to avoid
holding the threads in the queries executor.

Change-Id: I7a3ca0c0add90bd0d4f9601e647620c7dfd6258f