[vSphere] Select datastore for Disk locator spec
When a storage policy is chosen for boot disk, for setting the disk
locator spec a compatible datastore is needed. Since the storage policy
can span across datacenters, it was picking a wrong datastore belonging to
different datacenter for a chosen host from another datacenter.

Fixed this to pick the datastore belonging to the storage policy with in
the same datacenter as the host.

Change-Id: I5e872a66500d8e72ac4ab9ae56a7fed564b7f067