Remove queryTask.tenantLinks filter in the queries used by AWSEBSStorageEnumeration
Issue: Duplicate disks are created for the provisioned disks.

Cause: Because of queryTask.tenantLinks filter, the query is not fetching
the provisioned disks and as a result enumeration is creating duplicate
diskstates. Since scope constraints already adds boolean clauses for each
of the tenantLinks, queryTask.tenantLinks filter is not required.

Change-Id: Ice3c24667cc30d5bcca54e31fd2ad059d8bbc7be