Fix SSL certificate change notification to work with Postgres
This used to use continuous queries to detect certificate changes and
update the trust store on each node in the group. With Postgres
continuous queries only work on the owner node, so this is now changed
as follows:

1. The continuous query is used to detect certificate changes on the
owner node only.
2. The owner node then broadcasts a change notification that is received
by each node in the group. This is done through a new stateless helper
service.
3. The stateless service on each node calls the manager instance to
update the trust store on this node.

Change-Id: Ic5b3354a7bfac9928771b8ea74a8ffadb835b112